[Timer]: Timer is the best

+++ install_pack/ev3cleti_library/public/ev3_timer.h
@@ -4,7 +4,7 @@
 void Timer_reset();
 float Timer_get_time();
 void Timer_close();	
-void Timer_sleep_until(int sec);
-short Timer_count_from(int sec, int* currSec);
+void Timer_sleep_until(float sec);
+short Timer_count_from(float sec, float* currSec);
 
 #endif //!TIMER_H
